saturday was pretty good too
oh man, what a weekend. everything pulls down just like it should – the only thing is, can it get pushed back up? :) my thoughts on the matter at the moment: we're storing everything pulled and everything edited into two variables: presenceedit and presenceorig respectively. presenceedit is 'saved' (locally!) every time b_workload is called. it would stand to reason that, upon pressing the save button (and perhaps upon pressing a new cube?) that the two objects are compared against each other, and any changes are pushed up and dealt with accordingly. be sure that the presenceedit (at the position of currentlyediting) is updated to the latest 'version' of prescence before continuing. please note that at the moment, shapeshifter images are only being stored, and have not yet been set to be sprite-ified. okay I think that's it for now. I need sleep. oh yeah, how about new cubes / rows? :) that might be a little bit further down the road. have fun!