Merge, part 1
This is weird. Maybe even a little bit uncomfortable. .ajaxComplete() is kind of being a little bitch. Please note that no cubes are loading due to cubedescender not being enabled in editorinit. Oh yeah, I forgot – .ajaxComplete() fires every time an ajax event completes, including that of cubedescender. Perhaps we should make something that delivers the goods direct via PHP a la the front? Give it a shot…! <3