SO CLOSE
…and yet so far. but seriously, things are beginning to get wild.

after spending entirely too much time trying to figure out why the fuck GD doesn't work, breaking ject and practically reinstalling everything, it turns out it still doesn't fucking work. great. however, imagick on the server works perfectly! and its much more performant, so we'll use that. nice!

the problem here is that its a bitchwipe to install on a mac, so we're kinda stuck developing on the server to get it right…

…which sucks, because its flipping the image upside down for some strange reason. fix it, and I think we'll be

ROCK

FUCKING

SOLID

yeah? yeah!